Fix to recover a session
It happens a problem when a login was made and the session was expired
using Google Chrome in background, a property of saved session was
corrupted and we had problem to recover it. Was added an error treatment
in GetObjectFromBytes function of RedisUtility class, this error is
saved on project log.